check services for restricted by categories of the client
